add header component
